Use full snippet
