+++ docs/ai/quickstarts/snippets/mcp-server/.mcp/server.json
@@ -1,0 +1,27 @@
+{
+  "description": "<your description here>",
+  "name": "io.github.<your GitHub username here>/<your repo name>",
+  "packages": [
+    {
+      "registry_name": "nuget",
+      "name": "SampleMcpServer",
+      "version": "0.1.0-beta",
+      "package_arguments": [],
+      "environment_variables": [
+        {
+          "name": "WEATHER_CHOICES",
+          "description": "Comma separated list of weather descriptions to randomly select.",
+          "is_required": true,
+          "is_secret": false
+        }
+      ]
+    }
+  ],
+  "repository": {
+    "url": "https://github.com/<your GitHub username here>/<your repo name>",
+    "source": "github"
+  },
+  "version_detail": {
+    "version": "0.1.0-beta"
+  }
+}
